requirements.txt
Flask
qrcode[pil]
pyzbar
Pillow